Allow for POST requests to return a response
This change is mainly to solve the Google Apps script
HTTP client issue wherby you cannot send a GET request
with a body.  So this will allow endpoints like
/harold/deploy/get_salon_names to be able to be a POST
request endpoint so that it can have a body.  We need
the request to have a body because the body data is what
is signed by the HTTP client making the request.