Fix bugs for setting/removing deploy holds (#44)
* Fix bugs for setting/removing deploy holds

1. Fix for setting a hold

The value of the 'reason' input param for the 'hold' and 'hold_all'
method of the DeployMonitor class is a tuple and not a string.
The reason text is the first member of the tuple.

2. Fix hold/unhold HTTP request bug

The the code before this commit, we were incorrectly
trying to make a call to fetch a salon by calling a function
that returns the data in an asynchronous manner.  The reason
for this is that before we were trying to fetch the salon
and do some verification and return an error to the HTTP
client if the salon could not be found.  This caused a bug
when trying to set/remove a deploy hold using the HTTP endpoints.

With this fix, we simple blindly make the call to hold/unhold.
If the salon cannot be found the call simply does nothing.  This
is simpler even though any error is not communicated to the
client.

3. Update comment

* Reason string parsing fix